fix(engine): fix jssg codemod runs performance issue removing polling
