+++ crates/core/tests/engine_tests.rs
@@ -8,10 +8,18 @@ use std::collections::HashMap;
 use std::fs;
 use std::path::PathBuf;
 use std::process::Command;
+use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::{Arc, Mutex};
+use std::time::{Duration, Instant};
 use tempfile::TempDir;
+use tokio::sync::Notify;
 
-use butterflow_core::engine::{CapabilitiesData, Engine};
+use butterflow_core::engine::{
+    await_js_ast_grep_execution_task, build_js_ast_grep_idle_timeout_message, finish_unit_progress,
+    js_ast_grep_idle_timeout, record_output_progress, record_unit_progress,
+    select_shard_scan_eligible_files, should_manage_git_for_node, CapabilitiesData, Engine,
+    StepPhase, StepProgressState, UnitProgressState, JS_AST_GREP_IDLE_TIMEOUT_MS_DEFAULT,
+};
 use butterflow_core::structured_log::StructuredLogger;
 use butterflow_core::workflow_runtime::{WorkflowCommand, WorkflowSession};
 use butterflow_core::{
@@ -26,6 +34,7 @@ use butterflow_models::step::{
 use butterflow_models::strategy::Strategy;
 use butterflow_models::trigger::TriggerType;
 use codemod_llrt_capabilities::types::LlrtSupportedModules;
+use codemod_sandbox::sandbox::engine::CodemodOutput;
 
 use butterflow_models::{DiffOperation, FieldDiff, TaskDiff};
 use butterflow_state::local_adapter::LocalStateAdapter;
@@ -6093,3 +6102,206 @@ async fn test_expression_resolution_nonexistent_variable() {
 // TODO: test_cycle_detection_direct_cycle
 // TODO: test_find_cycle_in_chain
 // TODO: test_runtime_cycle_detection
+
+#[test]
+fn js_ast_grep_idle_timeout_uses_default_and_respects_env_override() {
+    let _guard = EnvVarGuard::unset("CODEMOD_JS_AST_GREP_IDLE_TIMEOUT_MS");
+    assert_eq!(
+        js_ast_grep_idle_timeout(),
+        Duration::from_millis(JS_AST_GREP_IDLE_TIMEOUT_MS_DEFAULT)
+    );
+
+    std::env::set_var("CODEMOD_JS_AST_GREP_IDLE_TIMEOUT_MS", "1234");
+    assert_eq!(js_ast_grep_idle_timeout(), Duration::from_millis(1234));
+}
+
+#[test]
+fn shard_scan_falls_back_to_selector_matches_when_dry_run_finds_no_edits() {
+    let eligible = select_shard_scan_eligible_files(
+        Vec::new(),
+        vec!["src/a.ts".to_string(), "src/b.ts".to_string()],
+    );
+
+    assert_eq!(eligible, vec!["src/a.ts", "src/b.ts"]);
+}
+
+#[test]
+fn shard_scan_prefers_modified_files_when_available() {
+    let eligible = select_shard_scan_eligible_files(
+        vec!["src/changed.ts".to_string()],
+        vec!["src/selector-only.ts".to_string()],
+    );
+
+    assert_eq!(eligible, vec!["src/changed.ts"]);
+}
+
+#[test]
+fn managed_git_mode_is_enabled_for_local_pull_request_nodes() {
+    let _guard = EnvVarGuard::unset("BUTTERFLOW_STATE_BACKEND");
+    let node = Node {
+        id: "apply-transforms".to_string(),
+        name: "Apply AST Transformations".to_string(),
+        description: None,
+        r#type: butterflow_models::node::NodeType::Automatic,
+        depends_on: vec![],
+        trigger: None,
+        strategy: None,
+        runtime: None,
+        steps: vec![],
+        env: HashMap::new(),
+        branch_name: None,
+        pull_request: Some(butterflow_models::step::PullRequestConfig {
+            title: "PR".to_string(),
+            body: None,
+            draft: Some(true),
+            base: None,
+        }),
+    };
+
+    assert!(should_manage_git_for_node(&node));
+}
+
+#[test]
+fn record_unit_progress_updates_global_and_active_units() {
+    let state = Arc::new(std::sync::Mutex::new(StepProgressState::new()));
+    let before = state.lock().unwrap().global_last_progress_at;
+
+    std::thread::sleep(Duration::from_millis(5));
+    record_unit_progress(&state, "src/example.ts", StepPhase::ExecutionStarted);
+
+    let snapshot = state.lock().unwrap();
+    assert_eq!(snapshot.global_phase, StepPhase::ExecutionStarted);
+    assert!(snapshot.global_last_progress_at > before);
+    let unit = snapshot.active_units.get("src/example.ts").unwrap();
+    assert_eq!(unit.phase, StepPhase::ExecutionStarted);
+    assert!(unit.last_progress_at > before);
+    assert!(snapshot.output_active_units.contains("src/example.ts"));
+}
+
+#[test]
+fn record_output_progress_refreshes_executing_units() {
+    let state = Arc::new(std::sync::Mutex::new(StepProgressState::new()));
+    record_unit_progress(&state, "src/example.ts", StepPhase::ExecutionStarted);
+    let before = state
+        .lock()
+        .unwrap()
+        .active_units
+        .get("src/example.ts")
+        .unwrap()
+        .last_progress_at;
+
+    std::thread::sleep(Duration::from_millis(5));
+    record_output_progress(&state);
+
+    let snapshot = state.lock().unwrap();
+    assert_eq!(snapshot.global_phase, StepPhase::Output);
+    let unit = snapshot.active_units.get("src/example.ts").unwrap();
+    assert_eq!(unit.phase, StepPhase::Output);
+    assert!(unit.last_progress_at > before);
+}
+
+#[test]
+fn finish_unit_progress_removes_active_unit() {
+    let state = Arc::new(std::sync::Mutex::new(StepProgressState::new()));
+    record_unit_progress(&state, "src/example.ts", StepPhase::ExecutionStarted);
+    finish_unit_progress(&state, "src/example.ts", StepPhase::ExecutionFinished);
+
+    let snapshot = state.lock().unwrap();
+    assert_eq!(snapshot.global_phase, StepPhase::ExecutionFinished);
+    assert!(!snapshot.active_units.contains_key("src/example.ts"));
+    assert!(!snapshot.output_active_units.contains("src/example.ts"));
+}
+
+#[test]
+fn build_idle_timeout_message_uses_stalest_active_unit() {
+    let now = Instant::now();
+    let mut state = StepProgressState::new();
+    state.global_last_progress_at = now - Duration::from_secs(90);
+    state.global_phase = StepPhase::Output;
+    state.active_units.insert(
+        "src/fresh.ts".to_string(),
+        UnitProgressState {
+            last_progress_at: now - Duration::from_secs(10),
+            phase: StepPhase::Output,
+        },
+    );
+    state.active_units.insert(
+        "src/stale.ts".to_string(),
+        UnitProgressState {
+            last_progress_at: now - Duration::from_secs(75),
+            phase: StepPhase::ExecutionStarted,
+        },
+    );
+
+    let message = build_js_ast_grep_idle_timeout_message(&state, Duration::from_secs(60));
+    assert!(message.contains("src/stale.ts"));
+    assert!(message.contains("execution started"));
+    assert!(message.contains("active units: 2"));
+}
+
+#[tokio::test]
+async fn await_js_ast_grep_execution_task_returns_idle_timeout_error() {
+    let progress_state = Arc::new(std::sync::Mutex::new(StepProgressState::new()));
+    record_unit_progress(
+        &progress_state,
+        "src/stalled.ts",
+        StepPhase::ExecutionStarted,
+    );
+    let idle_timed_out = Arc::new(AtomicBool::new(false));
+    let idle_notify = Arc::new(Notify::new());
+    let idle_failure_message = Arc::new(std::sync::Mutex::new(None::<String>));
+
+    let local = tokio::task::LocalSet::new();
+    let idle_timed_out_for_task = Arc::clone(&idle_timed_out);
+    let idle_notify_for_task = Arc::clone(&idle_notify);
+    let idle_failure_message_for_task = Arc::clone(&idle_failure_message);
+    let progress_state_for_task = Arc::clone(&progress_state);
+    let result = local
+            .run_until(async move {
+                let trigger = tokio::spawn({
+                    let idle_timed_out = Arc::clone(&idle_timed_out_for_task);
+                    let idle_notify = Arc::clone(&idle_notify_for_task);
+                    let idle_failure_message = Arc::clone(&idle_failure_message_for_task);
+                    async move {
+                        tokio::time::sleep(Duration::from_millis(10)).await;
+                        idle_timed_out.store(true, Ordering::Release);
+                        if let Ok(mut message) = idle_failure_message.lock() {
+                            *message = Some(
+                                "No progress observed for 1s while processing src/stalled.ts (execution started, active units: 1)"
+                                    .to_string(),
+                            );
+                        }
+                        idle_notify.notify_waiters();
+                    }
+                });
+
+                let execution_task = tokio::task::spawn_local(async move {
+                    futures_util::future::pending::<
+                        std::result::Result<
+                            CodemodOutput,
+                            codemod_sandbox::sandbox::errors::ExecutionError,
+                        >,
+                    >()
+                    .await
+                });
+
+                let result = await_js_ast_grep_execution_task(
+                    execution_task,
+                    idle_timed_out_for_task,
+                    idle_notify_for_task,
+                    idle_failure_message_for_task,
+                    progress_state_for_task,
+                    Duration::from_secs(1),
+                    "src/stalled.ts",
+                )
+                .await;
+                trigger.await.unwrap();
+                result
+            })
+            .await;
+
+    let error = result.expect_err("pending execution should time out");
+    let message = error.to_string();
+    assert!(message.contains("No progress observed"));
+    assert!(message.contains("src/stalled.ts"));
+}
